python/requirements.txt
pytest>=3.4.2
grpcio>=1.10.0
grpcio-tools>=1.10.0
